add working answer delete buttons and improve redirection

+++ public/js/question-view.js
@@ -9,21 +9,32 @@ document.addEventListener("DOMContentLoaded", async () => {
         answerBox.classList.add("card");
         let button;
         if (user.id === currentUserId){
-            button = `<form action="${window.location.href}/answers/${answer.id}" method="DELETE"><button type="submit" class="btn btn-danger">Delete answer</button></form>`;
+            button = `<button type="submit" class="btn btn-danger delete-answer-button" id=${answer.id}>Delete answer</button>`; // the "id" attribute of the "delete answer" button is the id number of the answer that clicking on it will delete
         } else button = "";
         //answerBox.classList.add("panel-default");
         answerBox.innerHTML = `<div class="card-body answer-box"><h5 class="card-title"><b>${user.username}</b> said:</h5><p>${message}</p>${button}</div>`;
         answersContainer.appendChild(answerBox);
     })
 
-    document.getElementById("delete-question-button").addEventListener('click', async e => {
+    if (document.querySelectorAll("#delete-question-button").length > 0) {
+        document.getElementById("delete-question-button").addEventListener('click', async e => {
+            try {
+                await fetch(window.location.href, { method: 'DELETE' });
+            } catch (e) {
+                console.log(e);
+            }
+            const homeURL = ['http:/', window.location.href.split('/')[2], 'questions'].join('/'); // redirects to the questions list without referencing "localhost" by modifying the current URL
+            window.location.href = homeURL;
+        })
+    }
+
+    document.querySelectorAll(".delete-answer-button").forEach(button => button.addEventListener("click", async event => {
         try {
-            await fetch(window.location.href, { method: 'DELETE' });
-        } catch (e) {
+            await fetch(`${window.location.href}/answers/${event.target.id}`, { method: "DELETE" });
+            window.location.reload(true);
+        } catch (e){
             console.log(e);
         }
-        console.log("Got here!");
-        const homeURL = ['http:/', window.location.href.split('/')[2], 'questions'].join('/'); // redirects to the questions list without referencing "localhost" by modifying the current URL
-        window.location.href = homeURL;
-    })
+    }))
+
 });
